refactor(gui): use new category api
